L23/DjangoProject/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
idna==3.4
Pillow==9.4.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
